split BlockMetaMetricCache
